ingestors/ical: parse calendars instead of invdividual events
Icalendar::Event.parse ignores anything in the iCalendar stream that
is not an event object, leading to events without timezone information
if this is provided in the calendar object instead of the event
object.

This commit moves to only using Icalendar::Calendar.parse similarly to
how Chris Want did previously[1], but without re-parsing the stream
with Icalendar::Event.parse, as Icalendar::Calendar.events already
provides the contained events.

Also, the parsing and timezone extraction are moved to
Ingestors::IcalIngestor, as these are not DRAC-specific needs.

A minor correctness issue is also fixed, with the default timezone
being made calendar-specific, otherwise events from an iCalendar
stream with multiple calendar objects would be extracted with the
timezone of the first calendar in the stream.

[1] fed1ea94fd4e19b2951d47ed617344b764e9f626