Read in the iCalendar uid as the event external_id,
This affects how uniqueness of events is calculated, and
in particular when dealing with events that have the same URL.
(To be ultra specific, Compute Ontario Colliquium series uses the
same URL for each event, so external_id is used to distinguish
events in the series).

See Event.check_exists for more about how this works.