A kludge to prevent a problematic issue with the way TeSS "fixes" keywords. As mentioned in a comment in Dictionary#best_match:
TODO: dive deeper into this because this is horrible and error-prone.
Elixir has this set to 0.3, bumping it higher breaks a lot of tests.
The issue here is that a keyword of "Research Commons" matches the
TargetAudience dictionary key "researcher".
in Event#fix_keywords, this keyword gets shifted to the target_audience
(which is dumb, random, and error-prone).
Ultimately, fix_keywords needs some better logic (or this function does).