A subclass of libcal to handle events in BC.
This isn't a great solution but will likely work for now.
The main issue is that libcal doesn't seem to have a timezone built in.
(Or if it does, it's hiding in documentation I don't have access to.)
The TeSS libcal ingestor sets the timezone to "Amsterdam", but otherwise
seems to work fine. So I've refactored that part of the ingestion into
a method, and override it in a subclass.

Also terrible: the output from libcal gives events in local time, so
having that timezone if crucial to store times in DB as UTC.
The cheap way: add a timezone of "Pacific Time (US & Canada)" to Event
and convert times.

I have a branch with some more correct ways to do things: add a default
timezone for a Source and also add option to convert times from local
to UTC. The code was getting complicated and not really great (really,
the entire ingestion pipeline needs a rewrite), so I'm going to take
the easy route for now. Likely if libcal gets more than two subclasses
it will be time to do the more complete solution.