Ensure exported iCalendar has an IANA timezone
Unfortunately, TeSS decided to use Rails' "curated list" of
timezones[1].

This list is the one used for verifying the timezone in the Rails
configuration. For this purpose is reasonable to limit the list to a
"meaningful subset of 134 zones", and even add some that are not in the
IANA list[2] (e.g., “Eastern Time (US & Canada)”).

The problem comes when you try to use this to represent iCalendar
dates, which can have any timezone in the IANA list. Exporting is also
a problem if a date is in a timezone not in the IANA list, the result
will depend on the implementation, most probably the date is assumed
to be in UTC.

This commit introduces a hack that maps Rails' "friendly" timezones
into pseudo-timezones that are in the IANA list.

[1] https://api.rubyonrails.org/v7.0/classes/ActiveSupport/TimeZone.html
[2] https://www.iana.org/time-zones