Moved "Register event" button to content provider page. Removed it from the events#index page. Shown only to people who control the content provider. (When clicked, the content provider is filled in the form.) Also, move the content provider form element higher up when manually entering an event.
This is in response to their being five events in Explora with no content provider.
There is no validation that a content provider must be present
(maybe there should be? We'll figure it out in time ...).
One of these was a UVic one that was the result of manually entering an event
(but not supplying a content provider).

We can assess whether this is a good move based on user feedback,
but as we add more content providers that want to manually enter events,
this will help them do the right thing.