This fixes a crash with the export options in the admin panel.
The problem is that translation for "activerecord.attributes.event.presence"
is expecting either a translation of the attribute name (not value), and nested values
are expected to handle pluralization of the attribute name. Instead, the column
values were stored here. Solution: move them out. Bonus: add a couple of helpers.