Exclude patterns for sources.
Currently supports patterns in the title or description (either strings or regex).
Adds a JSON column to the sources table to support this.

Concrete example: Compute Ontario will be adding events for the Compute Ontario Colloquia,
so we need to exclude these events when coming in from either SciNet or Sharcnet.